Update README.md
This is where I store all the examples and bits of code that I have learned in all my Python classes and doing projects.  
You are free to use any of these.  I wanted to consolidate all my tools in a single space so when I switch between laptops I always have access.